Return bytes for null and boolean in the encoder, not str
The tests hadn't caught this issue because we had a hack in
encode_call() that would force the encoder return values to be bytes if
they were str. But in cases where the boolean or null values were not
arguments to a call (e.g. if they were part of a nested data structure)
the encoding would fail with the error "Can't concat str to bytes"